changed what html is returned in holidays left handler

+++ holiday/holiday.go
@@ -11,7 +11,7 @@ type NextHoliday struct {
 	Name      string `json:"name"`
 	Date      string `json:"date"`
 	IsToday   bool   `json:"isToday"`
-	DaysUntil int8   `json:"daysUntil"`
+	DaysUntil int    `json:"daysUntil"`
 }
 
 type Holiday struct {
@@ -19,7 +19,7 @@ type Holiday struct {
 	Date string `json:"date"`
 }
 
-func NewNextHoliday(name string, date string, is_today bool, days_until int8) NextHoliday {
+func NewNextHoliday(name string, date string, is_today bool, days_until int) NextHoliday {
 	var next_holiday NextHoliday
 	next_holiday.Name = name
 	next_holiday.Date = date
@@ -57,11 +57,11 @@ func FindNextHoliday(holidays []Holiday) *Holiday {
 
 func GetRemainingHolidaysInYear(h *[]Holiday, year int) *[]Holiday {
 	var remainingHolidays []Holiday
-	today , _  := MakeDates(Holiday{})
+	today, _ := MakeDates(Holiday{})
 
 	for _, holiday := range *h {
 		_, holidayDate := MakeDates(holiday)
-		if  holidayDate.After(today) {
+		if holidayDate.After(today) {
 			remainingHolidays = append(remainingHolidays, holiday)
 		}
 	}
@@ -75,10 +75,10 @@ func (h Holiday) IsToday() bool {
 		holidayDate.Day() == currentDate.Day()
 }
 
-func (h Holiday) DaysUntil() int8 {
+func (h Holiday) DaysUntil() int {
 	currentDate, holidayDate := MakeDates(h)
 	daysUntil := math.Ceil(holidayDate.Sub(currentDate).Hours() / 24)
-	return int8(daysUntil)
+	return int(daysUntil)
 }
 
 func MakeDates(h Holiday) (time.Time, time.Time) {
